requirements.txt
Package                 Version
----------------------- ---------
amqp                    5.0.6
asgiref                 3.4.1
async-generator         1.10
attrs                   21.2.0
autopep8                1.5.7
beautifulsoup4          4.10.0
billiard                3.6.4.0
boto3                   1.20.10
botocore                1.23.10
celery                  5.2.1
certifi                 2021.10.8
cffi                    1.15.0
charset-normalizer      2.0.7
click                   8.0.3
click-didyoumean        0.3.0
click-plugins           1.1.1
click-repl              0.2.0
colorama                0.4.4
configparser            5.1.0
crayons                 0.4.0
cryptography            35.0.0
decorator               5.1.0
Django                  3.2.9
django-celery-beat      2.2.1
django-environ          0.8.1
django-redis            5.0.0
django-redis-cache      3.0.0
django-timezone-field   4.2.1
django-user-agents      0.4.0
djangorestframework     3.12.4
dnspython               2.1.0
elasticsearch           7.13.0
elasticsearch-dsl       7.4.0
fontawesome-free        5.15.4
gunicorn                20.1.0
h11                     0.12.0
idna                    3.3
install                 1.3.4
jmespath                0.10.0
jsonpickle              2.0.0
kombu                   5.2.2
logtail-python          0.1.2
lucene-querybuilder     0.2
lxml                    4.6.4
mongoengine             0.23.1
msgpack                 1.0.2
neo4j                   4.4.0
neo4j-driver            4.3.6
neobolt                 1.7.17
neomodel                4.0.7
outcome                 1.1.0
persisting-theory       0.2.1
Pillow                  8.4.0
pip                     21.2.4
prompt-toolkit          3.0.22
psycopg2                2.9.2
pycodestyle             2.8.0
pycparser               2.21
pymongo                 3.12.1
pyobjc-core             8.0
pyobjc-framework-Cocoa  8.0
pyobjc-framework-Quartz 8.0
pyOpenSSL               21.0.0
python-crontab          2.6.0
python-dateutil         2.8.2
pytz                    2021.3
PyYAML                  6.0
redis                   3.5.3
requests                2.26.0
s3transfer              0.5.0
selenium                4.0.0
setuptools              58.3.0
Shapely                 1.7.1
six                     1.16.0
sniffio                 1.2.0
sortedcontainers        2.4.0
soupsieve               2.3.1
sqlparse                0.2.4
toml                    0.10.2
trio                    0.19.0
trio-websocket          0.9.2
ua-parser               0.10.0
urllib3                 1.26.7
user-agents             2.2.0
vine                    5.0.0
wcwidth                 0.2.5
webdriver-manager       3.5.2
wheel                   0.37.0
whitenoise              5.3.0
wsproto                 1.0.0
